docs(packages): add READMEs and remixAssetServer middleware
hono-remix-middleware:
- split src/index.tsx into renderer.tsx + asset-server.ts; new index.tsx
  re-exports both
- new remixAssetServer(assets) middleware that wraps a remix/assets
  AssetServer instance as a Hono middleware (structural typing — no
  remix/assets import in the middleware itself, so Workers consumers
  don't accidentally pull in node:fs)
- README documents three deployment patterns side by side:
    1. Vite + Cloudflare Workers (vite-plugin-remix + assets binding)
    2. Vite + Node/Bun/Deno (vite-plugin-remix + Hono serveStatic)
    3. Native — no bundler, tsx/bun/deno (createAssetServer wrapped by
       remixAssetServer)

vite-plugin-remix:
- README walks through quickstart (vite.config.ts, entry.ts, Script),
  options reference, build output layout, the moduleUrl-as-lookup-key
  flow, and why builder.buildApp is left to the consumer
