+++ js/package/hono-remix-middleware/package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "exports": {
     ".": "./src/index.tsx",
+    "./asset-server": "./src/asset-server.ts",
     "./types": "./src/types.d.ts"
   },
   "peerDependencies": {
